+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -81,7 +81,7 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-23T09:49:01.865Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-24T23:05:31.950Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")