+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFN4436JPSWR6FY7QX5X58Z5"
+        "01KFS4386A1MM34M8SWVH3PP7T"
     )
     r"""The unique identifier of the data source"""